Adapt for user specific queries disallowing general wishlist access
