+++ src/query.rs
@@ -1,54 +1,20 @@
-use crate::{
-    base_connection::{BaseConnection, FindResultWrapper},
-    order_datatypes::WishlistOrderInput,
-    wishlist_connection::WishlistConnection,
-    Wishlist,
-};
+use crate::{user::User, Wishlist};
 use async_graphql::{Context, Error, Object, Result};
-use bson::Document;
+
 use bson::Uuid;
-use mongodb::{bson::doc, options::FindOptions, Collection, Database};
-use mongodb_cursor_pagination::{error::CursorError, FindResult, PaginatedCursor};
+use mongodb::{bson::doc, Collection, Database};
 
 /// Describes GraphQL wishlist queries.
 pub struct Query;
 
 #[Object]
 impl Query {
-    /// Retrieves all wishlists.
-    async fn wishlists<'a>(
+    /// Retrieve user with wishlists.
+    async fn user<'a>(
         &self,
-        ctx: &Context<'a>,
-        #[graphql(desc = "Describes that the `first` N wishlists should be retrieved.")]
-        first: Option<u32>,
-        #[graphql(desc = "Describes how many wishlists should be skipped at the beginning.")]
-        skip: Option<u64>,
-        #[graphql(desc = "Specifies the order in which wishlists are retrieved.")] order_by: Option<
-            WishlistOrderInput,
-        >,
-    ) -> Result<WishlistConnection> {
-        let db_client = ctx.data_unchecked::<Database>();
-        let collection: Collection<Wishlist> = db_client.collection::<Wishlist>("wishlists");
-        let wishlist_order = order_by.unwrap_or_default();
-        let sorting_doc = doc! {wishlist_order.field.unwrap_or_default().as_str(): i32::from(wishlist_order.direction.unwrap_or_default())};
-        let find_options = FindOptions::builder()
-            .skip(skip)
-            .limit(first.map(|v| i64::from(v)))
-            .sort(sorting_doc)
-            .build();
-        let document_collection = collection.clone_with_type::<Document>();
-        let maybe_find_results: Result<FindResult<Wishlist>, CursorError> =
-            PaginatedCursor::new(Some(find_options.clone()), None, None)
-                .find(&document_collection, None)
-                .await;
-        match maybe_find_results {
-            Ok(find_results) => {
-                let find_result_wrapper = FindResultWrapper(find_results);
-                let connection = Into::<BaseConnection<Wishlist>>::into(find_result_wrapper);
-                Ok(Into::<WishlistConnection>::into(connection))
-            }
-            Err(_) => return Err(Error::new("Retrieving wishlists failed in MongoDB.")),
-        }
+        #[graphql(desc = "UUID of user to retrieve.")] id: Uuid,
+    ) -> Result<User> {
+        Ok(User { _id: id })
     }
 
     /// Retrieves wishlist of specific id.
